Pulled LogEntry component out of log file. Added LogEntryContainer.
